Enum gdl_parser::Sentence

pub enum Sentence {
    PropSentence(Proposition),
    RelSentence(Relation),
}

A GDL sentence is like a Rule without a body. The two types of Sentences are Relations and Propositions

Variants

PropSentence

A proposition sentence

RelSentence

A relation sentence

Methods

impl Sentence

fn name(&self) -> &Constant

Returns the name of the sentence